requirements.txt
streamlit>=1.25.0
openai>=1.0.0
cohere>=4.26
pinecone-client>=2.2.2
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.1
scikit-learn>=1.3.0
numpy>=1.24.4
pandas>=2.0.3
